server/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.0
google-generativeai>=0.3.0
gunicorn>=20.1.0  # For production deployment
pytest>=7.0.0  # For testing 